requirements.txt
aiogram==3.0.*
aiohttp
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic